[MOBILE-3857] migrate config
